# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bmstu Open IT Platform",
        "title": "История и Культура 26. Расчет бюджета создания постройки/произведения с учетом исторического индекса цен",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/",
    "paths": {
        "/api/add_to_estimate/{id}": {
            "post": {
                "description": "Добавляет указанную услугу в текущую черновую заявку (корзину) пользователя. Если черновика нет, он создается.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "estimates"
                ],
                "summary": "Добавить услугу в черновую заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исторической услуги для добавления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном добавлении и информация о заявке",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверный ID параметра",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера (например, при создании черновика)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/create_object": {
            "post": {
                "description": "Создает новую запись исторической услуги в базе данных.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Создать новую историческую услугу",
                "parameters": [
                    {
                        "description": "Данные для создания услуги",
                        "name": "Historical_service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Historical_service"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно созданная услуга",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON или неверные данные",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/historical_object/{id}": {
            "get": {
                "description": "Возвращает полную информацию об исторической услуге по её ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Получить услугу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исторической услуги (объекта)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверный ID параметра",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Объект не найден",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет существующую услугу по её ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Обновить историческую услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исторической услуги для обновления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления услуги",
                        "name": "Historical_service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Historical_service"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно обновленная услуга",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверный ID или некорректный JSON",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет услугу из базы данных по её ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Удалить историческую услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исторической услуги для удаления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверный ID параметра",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/historical_object/{id}/image": {
            "post": {
                "description": "Загружает изображение для конкретной исторической услуги. Использует multipart/form-data.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Загрузить изображение для услуги",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исторической услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл изображения",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешной загрузке",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверный ID или файл не предоставлен",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера при обработке файла",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/historical_objects": {
            "get": {
                "description": "Возвращает список всех исторических услуг. Может быть отфильтрован по названию.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Получить список услуг",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название услуги (частичное совпадение)",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список исторических услуг",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ds.Historical_service": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dataSource": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "historicalPeriod": {
                    "type": "string"
                },
                "historicalRegion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageURL": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "priceUSD": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        }
    }
}